# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:8077",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [],
  "paths": {
    "/users/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for a user to signup",
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "User email",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "antonio@gmail.com"
                }
              }
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "User password",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1234
                },
                "token": {
                  "type": "string",
                  "example": "token_string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "406": {
            "description": "Not Acceptable"
          }
        }
      }
    },
    "/users/confirmation/{email}/{token}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for a user to signup",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User email",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "antonio@gmail.com"
                }
              }
            }
          },
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Verification token",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "verification_token"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "User is verified."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Your verification link may have expired. Please use 'resend' endpoint to resend the verification link. Check API documentation for endpoint details."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "We were unable to find a user for this verification. Please SignUp!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "[Database error]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/confirmation/resend": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for a user to signup",
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "User email",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "example": {
                  "type": "string",
                  "example": "antonio@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "This account has been already verified. Please log in."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "A verification email has been sent to [user.email]. It will be expire after one day. If you not get verification email, use 'resend' endpoint. Check API documentation for endpoint details."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "We were unable to find a user with that email. Make sure your Email is correct!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for a user to login. Will get a token back if successful.",
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "User email",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "antonio@gmail.com"
                }
              }
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "User password",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "example"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "token_string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Error: Unable to login"
          }
        }
      }
    },
...
    "/users/liked_recipes": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "List the recipe that user liked",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "recipe_name1"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Recipes were not found + [error]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/liked_ingredient_frequencies": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint for a user to list the ingredient frequency of liked meals",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "ingredient1": {
                  "type": "number",
                  "example": 5
                },
                "ingredient2": {
                  "type": "number",
                  "example": 2
                },
                "ingredient3": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "example": "Database error + [error]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}